+++ CHANGELOG.md
@@ -26,16 +26,17 @@
 - **tower:** added code block marshaler pattern
 - **tower-discord:** implemented file upload native discord
 - **tower-discord:** added thread id in metadata
+- **tower-http:** added unit test
 - **tower-http:** added client logger
 - **tower-http:** added respond body logger
 - **tower-http:** added logging middleware
 - **tower-http:** added logging middleware to respond error
 - **tower-http:** added unit test
-- **tower-http:** added unit test
-- **tower-http:** major bug fixes with compressions
 - **tower-http:** added response capturer
 - **tower-http:** unit tested respond ok
 - **tower-http:** unit tested respond ok with http no body
+- **tower-http:** major bug fixes with compressions
+- **tower-http-gzip:** skip compression on data too small
 
 ### Miscellaneous
 - synchronize go mod with new tag on tower